CMake code updates
This eliminates the added 4th number field on the library version
(so it matches the MUMPS version). The NAG compiler doesn't recognize
a .so library with 4-number version as a library.

This mucks around with the target_link_libraries, getting rid of the
dependence on MPI::MPI_C and using MPI::MPI_Fortran alone. This appears
to be responsible for resolving the bad, gcc-compatible, link line
argument that was being fed to nagfor. MUMPS only uses the mpif.h,
except in the libseq code which is only built in serial (we don't).
The one remaining problem is that the complex128 test (part of the
mumps cmake wrapper project, not MUMPS itself) is C code and it
fails to link because gcc is being fed nagfor-compatible link line
argument. Possibly may be fixed by using Fortran as the link language
for that test executable. But this is sidestepped by turning off
the building of the tests.

This also eliminates using C as the linker language that was added
before; the previous changes make this possible.